requirements.txt
opencv-python
pillow
pandas
numpy
mysql-connector-python
Pillow
tk

